+++ astropy/time/iers/setup_package.py
@@ -1,0 +1,7 @@
+def get_package_data():
+    return {
+        'astropy.time.iers': [
+            'data/ReadMe.eopc04_IAU2000',
+            'data/ReadMe.finals2000A',
+            'data/eopc04_IAU2000.62-now.gz']
+        }
