+++ astropy/utils/setup_package.py
@@ -23,8 +23,4 @@ def get_package_data():
             'data/*.dat.bz2',
             'data/*.txt',
             'data/.hidden_file.txt'],
-        'astropy.utils.iers': [
-            'data/ReadMe.eopc04_IAU2000',
-            'data/ReadMe.finals2000A',
-            'data/eopc04_IAU2000.62-now.gz']
     }
